Relocate poll_interval, max_poll_attempts and socket_timeout to config_mixin
